abstract common models
