+++ flameshow/main.py
@@ -6,7 +6,8 @@ import click
 from click.exceptions import UsageError
 
 
-from flameshow.parser import ProfileParser
+from flameshow.pprof_parser.parser import ProfileParser
+from flameshow.pprof_parser.parser import parse_profile
 from flameshow.render import FlameGraphApp
 from flameshow.pprof_parser import parse_golang_profile
 from flameshow.const import MAX_RENDER_DEPTH, MIN_RENDER_DEPTH
@@ -32,36 +33,31 @@ def setup_log(enabled, level, loglocation):
 LOG_LEVEL = {0: logging.CRITICAL, 1: logging.WARNING, 2: logging.INFO, 3: logging.DEBUG}
 
 
-def run_app(verbose, log_to, format, profile, _debug_exit_after_rednder):
+def run_app(verbose, log_to, format, profile_f, _debug_exit_after_rednder):
     log_level = LOG_LEVEL[verbose]
     setup_log(log_to is not None, log_level, log_to)
 
     logger.info(f"run app, {format=}")
 
     t0 = time.time()
-    profile_data = profile.read()
+    profile_data = profile_f.read()
     profile_dict = parse_golang_profile(profile_data)
     t01 = time.time()
-    logger.info("Read golang profile, took %.2fs", t01 - t0)
+    logger.info("Read golang profile_f, took %.1fs", t01 - t0)
 
     if format == "json":
         click.echo(json.dumps(profile_dict, indent=2, sort_keys=True))
     elif format == "flamegraph":
-        parser_obj = ProfileParser(filename=profile.name)
-        t1 = time.time()
-        profile = parser_obj.parse(profile_dict)
-        logger.info("Max depth: %s", parser_obj.highest)
-        t2 = time.time()
-        logger.info("parse json files cost %s s", t2 - t1)
+        profile = parse_profile(profile_dict, profile_f.name)
 
-        render_depth = MAX_RENDER_DEPTH - int(parser_obj.total_sample / 100)
+        render_depth = MAX_RENDER_DEPTH - int(profile.total_sample / 100)
         # limit to 3 - 10
         render_depth = max(MIN_RENDER_DEPTH, render_depth)
         render_depth = min(MAX_RENDER_DEPTH, render_depth)
 
         logger.info(
             "Start to render app, total samples=%d, render_depth=%d",
-            parser_obj.total_sample,
+            profile.total_sample,
             render_depth,
         )
