+++ flameshow/pprof_parser/__init__.py
@@ -1,3 +1,4 @@
 from .pprof_parser import parse_golang_profile
+from .parser import parse_profile
 
-__all__ = ["parse_golang_profile"]
+__all__ = ["parse_golang_profile", "parse_profile"]
